added login_required to admin views and statement to check if user has permission

+++ admin_actions/views.py
@@ -3,59 +3,72 @@ from .forms import ProductForm
 from products.models import Product
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
 
 
+@login_required
 def add_product(request):
     """ Add a product to the store """
-    if request.method == 'POST':
-        form = ProductForm(request.POST, request.FILES)
-        if form.is_valid():
-            form.save()
-            messages.success(request, 'Successfully added product!')
-            return redirect(reverse('add_product'))
+    if request.user.is_superuser:
+        if request.method == 'POST':
+            form = ProductForm(request.POST, request.FILES)
+            if form.is_valid():
+                form.save()
+                messages.success(request, 'Successfully added product!')
+                return redirect(reverse('add_product'))
+            else:
+                messages.error(request, 'Failed to add product. Please ensure the form is valid.')
         else:
-            messages.error(request, 'Failed to add product. Please ensure the form is valid.')
+            form = ProductForm()
+            
+        template = 'admin_actions/add_product.html'
+        context = {
+            'form': form,
+        }
+
+        return render(request, template, context)
     else:
-        form = ProductForm()
-        
-    template = 'admin_actions/add_product.html'
-    context = {
-        'form': form,
-    }
-
-    return render(request, template, context)
+        raise PermissionDenied
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit an existing product """
 
     # Get the product from the id passed as an argument
-    product = get_object_or_404(Product, pk=product_id)
-    
-    if request.method == 'POST':
-        form = ProductForm(request.POST, request.FILES, instance=product)
-        if form.is_valid():
-            form.save()
-            messages.success(request, 'Successfully updated product!')
-            return redirect(reverse('product_detail', args=[product.id]))
+    if request.user.is_superuser:
+        product = get_object_or_404(Product, pk=product_id)
+        
+        if request.method == 'POST':
+            form = ProductForm(request.POST, request.FILES, instance=product)
+            if form.is_valid():
+                form.save()
+                messages.success(request, 'Successfully updated product!')
+                return redirect(reverse('product_detail', args=[product.id]))
+            else:
+                messages.error(request, 'Failed to update product. Please ensure the form is valid.')
         else:
-            messages.error(request, 'Failed to update product. Please ensure the form is valid.')
+            form = ProductForm(instance=product)
+            messages.info(request, f'You are editing {product.name}')
+
+        template = 'admin_actions/edit_product.html'
+        context = {
+            'form': form,
+            'product': product,
+        }
+
+        return render(request, template, context)
     else:
-        form = ProductForm(instance=product)
-        messages.info(request, f'You are editing {product.name}')
-
-    template = 'admin_actions/edit_product.html'
-    context = {
-        'form': form,
-        'product': product,
-    }
-
-    return render(request, template, context)
+        raise PermissionDenied
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
-    product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, 'Product deleted!')
-    return redirect(reverse('products'))
+    if request.user.is_superuser:
+        product = get_object_or_404(Product, pk=product_id)
+        product.delete()
+        messages.success(request, 'Product deleted!')
+        return redirect(reverse('products'))
+    else:
+        raise PermissionDenied
